requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
jinja2==3.1.2
markdown==3.4.4
aiosmtplib==1.1.6
python-dateutil==2.8.2
pytz==2024.1
SQLAlchemy==2.0.22
python-dotenv==1.0.0
requests==2.31.0
